docker_mpt/requirements.txt
telethon>=1.28.5
sqlalchemy>=2.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiofiles>=23.2.1
alembic>=1.12.0
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0 
python_socks